Added Scene Manager, Scene, and Scene1 classes integrated into main file.
